pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trinity-atlassian-cli"
version = "0.1.0"
description = "Unified Atlassian CLI for Jira, Confluence, and Bitbucket — agent-native, JSON-first"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Ethan Drower", email = "ethan@citemed.com"}
]
keywords = ["atlassian", "jira", "confluence", "bitbucket", "cli", "agent", "ai", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Tools",
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "requests>=2.28.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
trinity = "trinity.cli:main"

[project.urls]
Homepage = "https://github.com/ethandrower/trinity-atlassian-cli"
Repository = "https://github.com/ethandrower/trinity-atlassian-cli.git"
Issues = "https://github.com/ethandrower/trinity-atlassian-cli/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.coverage.run]
source = ["src/trinity"]
omit = ["*/tests/*", "*/test_*"]
